feat: Rename inferred-spans to span-inferrer, as it exports a class & follows command pattern. Update imports
