+++ src/trace/span-inferrer.ts
